Alterações exercicios Aula03

+++ Aula03/Exercicio02/Program.cs
@@ -11,14 +11,14 @@ namespace Exercicio02
         static void Main(string[] args)
         {
             //2. Converter a temperatura em graus Fahrenheit para graus Celsius.
-            double fahrenheit;
-            double celsius;
+            float fahrenheit;
+            float celsius;
 
             Console.WriteLine("Converter Fahrenheit para Celsius\n");
             Console.WriteLine("Digite a quantidade de graus Fahrenheit");
-            fahrenheit = double.Parse(Console.ReadLine());
+            fahrenheit = float.Parse(Console.ReadLine());
 
-            celsius = (fahrenheit-32) / 1.8;
+            celsius = (fahrenheit-32) / 1.8f;
 
             Console.WriteLine($"O resultado é {Math.Round(celsius,1)}º Celsius");
             Console.ReadKey();
